requirements.txt
torch
torchvision
torchaudio
transformers==4.40.0
datasets
huggingface-hub
sentencepiece
accelerate
Pillow
timm
opencv-python
open-clip-torch
ftfy
regex
langchain
langchain-community
langchain-huggingface
numpy
pandas
scikit-learn
tqdm
requests
streamlit